Secured API Routes

+++ src/pages/api/posts/[id].ts
@@ -2,6 +2,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 import Posts from '../../../models/Posts'
 import dbConnect from '../../../utils/dbConnect'
+import { getToken } from "next-auth/jwt"
 
 dbConnect()
 
@@ -9,6 +10,7 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET })
   const { id } = req.query
   switch (req.method){
 
@@ -25,6 +27,7 @@ export default async function handler(
 
     case 'PUT': {
       try {
+        if (!token) return res.status(401).json({status: false}) 
         const post = await Posts.findByIdAndUpdate(id, req.body, {
           new: true,
           runValidators: true
@@ -39,6 +42,7 @@ export default async function handler(
 
     case 'DELETE': {
       try {
+        if (!token) return res.status(401).json({status: false}) 
         const deletedPost = await Posts.deleteOne({_id: id})
         if (!deletedPost) return res.status(400).json({status: false})
         res.status(201).json({status: true})
